+++ src/cz/tul/dic/data/task/TaskParameter.java
@@ -29,7 +29,7 @@ public enum TaskParameter {
     ROUND_LIMITS(int[].class),
     DISPLACEMENT_CALCULATION(DisplacementCalculation.class),
     STRAIN_ESTIMATION_METHOD(StrainEstimationMethod.class),
-    STRAIN_ESTIMATION_PARAM(Integer.class),
+    STRAIN_ESTIMATION_PARAM(Double.class),
     LOCAL_SEARCH_PARAM(Integer.class),
     ;
